+++ R/nested_laplace_joint_multi.R
@@ -722,7 +722,8 @@
                                          force_sparse, cell_coupling,
                                          diagnose_k = TRUE, k_samples = 200L,
                                          n_threads_outer = 1L, proposal = NULL,
-                                         pareto_k_by_arm = FALSE) {
+                                         pareto_k_by_arm = FALSE,
+                                         k_batches = 1L) {
     res$pareto_k        <- NA_real_
     res$pareto_k_is_ess <- NA_real_
     res$pareto_k_scope  <- "outer (hyperparameter) Gaussian proposal"
@@ -774,10 +775,11 @@
     refit <- .joint_make_diag_refit(res, solve_fn, modal_theta, knobs)
     arm_axes <- if (isTRUE(pareto_k_by_arm)) .joint_pareto_arm_axes(res) else NULL
     kd <- .joint_pareto_k(res, refit, k_samples, proposal = proposal,
-                          arm_axes = arm_axes)
+                          arm_axes = arm_axes, k_batches = k_batches)
     res$pareto_k        <- kd$pareto_k
     res$pareto_k_is_ess <- kd$is_ess
     res$pareto_k_proposal_source <- kd$proposal_source
+    res <- .joint_attach_batch_k(res, kd)
     res <- .joint_attach_by_arm_k(res, kd)
     res
 }
@@ -801,6 +803,7 @@
                                   k_samples = 200L,
                                   pareto_k_by_arm = FALSE,
                                   pareto_k_threads = NULL,
+                                  k_batches = 1L,
                                   inner_refresh = 1L,
                                   integration = "auto",
                                   timer = NULL) {
@@ -1149,7 +1152,8 @@
                                         k_samples  = k_samples,
                                         n_threads_outer = k_to,
                                         proposal = ccd_proposal,
-                                        pareto_k_by_arm = pareto_k_by_arm)
+                                        pareto_k_by_arm = pareto_k_by_arm,
+                                        k_batches = k_batches)
     tm$mark("diagnostics")
     res$timing <- tm$timing()
     .finalize_fit(res, backend = "nested_laplace_joint",